Fix pytest skipping, use pytest markers, fix tests (#3)
* Fix pytest skipping, use pytest markers, fix tests

Fixing some tests exposed an issue with pytest integration where the
wrapper was never being applied as it was being used on a copy of the
test function, not the original.

In debugging this, I decided to move us from custom function wrappers to
pytest markers. The output is nicer and we'll play nicer with other
pytest plugins and tooling.

I discovered the original pytest bug by fixing some integration tests
that incorrectly passed -- pytest was not being installed or invoked for
the pytest integration tests, so the result seemed like "correctly
skipping tests" when it was actually "no tests were run". 😅

* CR comments & add calls to start_test, end_test

* Bump version and update author to generic YB email